requirements.txt
python-dotenv 
pandas 
numpy
seaborn
matplotlib
xgboost
lightgbm
scikit-learn
jinja2>=3.1
ipywidgets>=8.0
pymongo
certifi
pyaml
fastapi 
uvicorn
##-e .